+++ client/src/utils/dataFetcher.js
@@ -269,6 +269,15 @@ const getVersionContent = async (documentId, versionNumber) => {
   }
 };
 
+const getColorForUser = (userId) => {
+  const colors = ["#fa3", "#39f", "#2e2", "#f39", "#c83"];
+  let hash = 0;
+  for (let i = 0; i < userId.length; i++) {
+    hash = userId.charCodeAt(i) + ((hash << 5) - hash);
+  }
+  return colors[Math.abs(hash) % colors.length];
+}
+
 export {
   getUserData,
   getUserDocuments,
@@ -283,4 +292,5 @@ export {
   revertToVersion,
   updateDocumentTitle,
   getVersionContent,
+  getColorForUser
 };
